Renamed and refined convert_text_json_wrapper, tests refined (delete temp files afterwards), setup venv

+++ test/tests.py
@@ -1,7 +1,8 @@
 import os
 from unittest import TestCase
 
-from src.main import sane_file_extensions, sane_file_naming_schema, Media, write_json, read_json, txt_to_html_to_json
+from src.convert_txt_to_html_to_json import convert_txt_to_html_to_json
+from src.main import sane_file_extensions, sane_file_naming_schema, Media, write_json, read_json
 
 
 def create_media_object(**kwargs):
@@ -101,16 +102,24 @@ class FileNamingTests(TestCase):
         self.assertTrue(write_json(['2019-05-27_Sonnenuntergang in Petershausen.jpg'], "test.json"))
         print("")
 
+    def tearDown(self):
+        if os.path.isfile("test.json"):
+            os.remove("test.json")
 
-class TxtToJsonTests(TestCase):
+
+class ConvertTxtToHtmlToJsonTests(TestCase):
 
     def test_convert_txt_to_html_to_json_positive(self):
         print("--> ", self._testMethodName)
         text = "Lorem ipsum dolor sit amet.\nNeue Zeile.\n    Zeile mit 4 Leerzeichen\n   Zeile mit drei Leerzeichen."
         with open("test.txt", "w") as file:
             file.write(text)
-        txt_to_html_to_json("test.txt", "test.json")
+        convert_txt_to_html_to_json("test.txt", "test.json")
         self.assertTrue(read_json("test.json"))
-        os.remove("test.txt")
-        os.remove("test.json")
         print("")
+
+    def tearDown(self):
+        if os.path.isfile("test.txt"):
+            os.remove("test.txt")
+        if os.path.isfile("test.json"):
+            os.remove("test.json")
